+++ lib/stellwerk/commands/check.rb
@@ -1,11 +1,9 @@
 require "pathname"
-require "reference_extractor"
 require "zeitwerk"
-require "parallel"
 
 require "stellwerk/config"
 require "stellwerk/printer"
-require "stellwerk/source_file_collector"
+require "stellwerk/graph"
 
 module Stellwerk
   module Commands
@@ -19,20 +17,8 @@ module Stellwerk
         puts "running check..."
         config = Stellwerk::Config.new(@root_path.join("stellwerk.yml"))
 
-        # build graph using reference_extractor
-        extractor = ReferenceExtractor::Extractor.new(
-          autoloaders: @autoloaders,
-          root_path: @root_path
-        )
-
-        all_files = Stellwerk::SourceFileCollector.new(@root_path).call
-        puts "collected #{all_files.length} files"
-
-        before = Time.now
-        edgelist = Parallel.flat_map(all_files, in_processes: Parallel.processor_count - 1) do |file|
-          extractor.references_from_file(file)
-        end
-        puts "extracted #{edgelist.length} references in #{(Time.now - before).round(2)} seconds"
+        edgelist = Stellwerk::Graph.new(@root_path, autoloaders: @autoloaders).build
+        puts "extracted #{edgelist.length} references"
 
         # check rules against graph, collect violations
         puts "checking rules..."